do not append variable name to calculation names #268
Changes to calc/base allowing function classes to overload how
the derived variable alias is determined. The target variable
alias is not appended in the case of a univariate calculation over
one variable. It is appended if the calculation is applied to multiple
variables. Tests were updated for the new case.

Closes #268.