implement rtree spatial index #291
If the module "rtree" is available, use it to construct real-time
spatial indexes. The old spatial index module has been replaced. The
spatial operations without "rtree" are implemented as simply as
possible and can become quite slow for complicated GIS operations.
There are also changes to the environment module that will detect
the presence of a module and update the environment variable associated
with the input.

closes #291