variable attributes written after value set #350
When creating a netCDF file, variable attributes are now written before
the value is set. This way scales and offsets may be appropriately
applied by netCDF4-python. Also, the keys “scale_value” and “add_offset”
are removed from the attributes dictionary when units are updated. The
data type of a conformed units value is no also appropriately updated to
ensure it matches the output from cfunits-python.

closes #350