seasonal time region subsets choking on mid-year origin time series #352
A check was added to the extraction of time regions for the purposes of
seasonal aggregation. This check ensures each generated time region is
valid for the time series used to construct it. This issue only occurred
with time-series beginning in mid-year where certain month-year
combinations were not actually in the time series.