support unmelted tabular output format #353
The default iteration method for creating tabular output formats is
now unmelted. Variables aliases are each given their own column. An
environment variable was added to allow the melted parameter to be
globally set. Multiple request datasets still require a melted form
which will be adjusted automatically with a warning.