capture warnings appropriately #164
Added a new environment variable to control suppression of warnings. The
built-in logging module warning capture system is now used. Suppressed
warnings will still be logged to file or stream depending on the logging
configuration. Due to the high volume of warnings related to file
metadata, application-specific warnings are suppressed by default.