automatically set GDAL_DATA variable if not present #358
``GDAL_DATA`` is now set in the ``environment`` module. This requires
all ``osgeo`` imports to only import from ``environment``. The configuration
is quirky, and once ``osgeo`` is imported, it is not possible to change
the configuration. A warning is issued if ``GDAL_DATA`` is not set advising
the user to set the environment variable for improved load
performance.

Also upgraded the following packages:
 * numpy==1.9.2
 * fiona==1.4.8
 * shapefly=1.5.6

Upgrading ``numpy`` resulted in some precision errors requiring test
values to be modified. No issues with the other packages.

closes #358